requirements.txt
setuptools~=49.2.1
opencv-python~=4.5.1.48
numpy~=1.20.3
scipy~=1.6.1
scikit-image~=0.19.2
torch~=1.8.0+cu111
geojson~=2.5.0
matplotlib~=3.3.4
Pillow~=8.4.0
torchvision~=0.9.0+cu111
imageio~=2.9.0